+++ src/main/java/capdb/design/repository/DateCourseRepository.java
@@ -3,6 +3,6 @@ package capdb.design.repository;
 import capdb.design.domain.DateCourse;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DateCourseRepository extends JpaRepository<DateCourse, Integer> {
-    //여기서 구현...?
+public interface DateCourseRepository extends JpaRepository<DateCourse, Long> {
+
 }
